django-oidc/requirements.txt
mozilla-django-oidc==2.0.0
